+++ src/pages/Account/styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+    imageContainer: {
+        justifyContent: "flex-start",
+        alignItems: "center",
+        paddingTop: 30,
+    },
+    userPhoto: {
+        width: "40%",
+        height: "50%",
+    },
+    userName: {
+        fontSize: 25,
+        color: "#333",
+        paddingTop: 20,
+        fontWeight: "bold",
+    },
+});