+++ src/pages/Help/index.js
@@ -1,10 +1,11 @@
 import React from "react";
 import { View, Text } from "react-native";
 import Header from "../Header";
+import globalStyles from "../globalStyle/globalStyles";
 
 export default function Help() {
     return (
-        <View>
+        <View style={globalStyles.container}>
             <Header />
             <Text>HELP</Text>
         </View>